Refactor `common::executor` by moving to a separate module
