+++ mm2src/coins/lightning.rs
@@ -32,8 +32,9 @@ use bitcoin_hashes::sha256::Hash as Sha256;
 use bitcrypto::dhash256;
 use bitcrypto::ChecksumType;
 use chain::TransactionOutput;
+use common::executor::spawn_abortable;
 use common::log::{error, LogOnError, LogState};
-use common::{async_blocking, calc_total_pages, log, now_ms, spawn_abortable, ten, PagingOptionsEnum};
+use common::{async_blocking, calc_total_pages, log, now_ms, ten, PagingOptionsEnum};
 use futures::{FutureExt, TryFutureExt};
 use futures01::Future;
 use keys::{hash::H256, AddressHashEnum, CompactSignature, KeyPair, Private, Public};
