+++ mm2src/coins/lightning/ln_events.rs
@@ -5,9 +5,9 @@ use crate::lightning::ln_sql::SqliteLightningDB;
 use bitcoin::blockdata::script::Script;
 use bitcoin::blockdata::transaction::Transaction;
 use bitcoin::consensus::encode::serialize_hex;
-use common::executor::Timer;
+use common::executor::{spawn_abortable, Timer};
 use common::log::{error, info};
-use common::{now_ms, spawn_abortable};
+use common::now_ms;
 use core::time::Duration;
 use futures::compat::Future01CompatExt;
 use lightning::chain::chaininterface::{ConfirmationTarget, FeeEstimator};
