+++ mm2src/coins/lightning/ln_platform.rs
@@ -13,9 +13,8 @@ use bitcoin::blockdata::transaction::Transaction;
 use bitcoin::consensus::encode::{deserialize, serialize_hex};
 use bitcoin::hash_types::{BlockHash, TxMerkleNode, Txid};
 use bitcoin_hashes::{sha256d, Hash};
-use common::executor::Timer;
+use common::executor::{spawn_abortable, AbortOnDropHandle, Timer};
 use common::log::{debug, error, info};
-use common::AbortOnDropHandle;
 use futures::compat::Future01CompatExt;
 use futures::future::join_all;
 use keys::hash::H256;
