+++ mm2src/mm2_main/src/lp_swap.rs
@@ -62,9 +62,9 @@ use async_std::sync as async_std_sync;
 use coins::{lp_coinfind, MmCoinEnum, TradeFee, TransactionEnum};
 use common::log::{debug, warn};
 use common::{bits256, calc_total_pages,
-             executor::{spawn, Timer},
+             executor::{spawn, spawn_abortable, AbortOnDropHandle, Timer},
              log::{error, info},
-             now_ms, spawn_abortable, var, AbortOnDropHandle, PagingOptions};
+             now_ms, var, PagingOptions};
 use derive_more::Display;
 use http::Response;
 use mm2_core::mm_ctx::{from_ctx, MmArc};
